Split config into multiple files
